backend/requirements.txt
asgiref==3.3.1
click==7.1.2
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-extensions==3.1.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.7.5
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
Werkzeug==1.0.1
whitenoise==5.2.0
